+++ Week-12/RestApiDevelopment/src/main/java/dev/patika/RestApiDevelopment/business/abstracts/IBookService.java
@@ -1,14 +1,13 @@
 package dev.patika.RestApiDevelopment.business.abstracts;
 
-import dev.patika.RestApiDevelopment.entities.Author;
 import dev.patika.RestApiDevelopment.entities.Book;
 
 import java.util.List;
 
 public interface IBookService {
-    Book getById(int id);
+    Book getById(Long id);
     Book save(Book book);
-    void delete(int id);
+    void delete(Long id);
     Book update(Book book);
     List<Book> findAll();
 }
