+++ Week-12/RestApiDevelopment/src/main/java/dev/patika/RestApiDevelopment/dao/BookBorrowingRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookBorrowingRepo extends JpaRepository<BookBorrowing, Integer> {
+public interface BookBorrowingRepo extends JpaRepository<BookBorrowing, Long> {
 }
